change script that is being run in dvc pipeline
