+++ pipelines/bertmesh/dvc.lock
@@ -1,7 +1,7 @@
 schema: '2.0'
 stages:
   train:
-    cmd: python ../../scripts/train_bertmesh_script.py --model_key Wellcome/WellcomeBertMesh
+    cmd: python ../../grants_tagger_light/training/train.py --model_key Wellcome/WellcomeBertMesh
       --data_path ../../data/raw/allMeSH_2021.json --output_dir ../../bertmesh_outs/pipeline_test
       --max_samples 10 --per_device_train_batch_size 4 --per_device_eval_batch_size
       4
@@ -14,6 +14,6 @@ stages:
         BertMeshTrainingArguments: {}
     outs:
     - path: ../../bertmesh_outs/pipeline_test/best
-      md5: 734a4e2eba3208449473dfa6cbee4683.dir
+      md5: dfc280687bcbda1eebfcf1ea68662d89.dir
       size: 649161367
       nfiles: 3
